refactoring: /question/save 엔드포인트 동작 확인
- 사용자 엔티티의 아이디가 없음으로 인해 외래키 조건에 부합하여 오류 발생
- 스프링 문제가 아닌 데이터베이스 문제
- 사용자 엔티티 수정 (udid 추가 및 rid 삭제)

+++ src/main/java/com/example/androidserver/web/Controller/UserController.java
@@ -31,7 +31,7 @@ public class UserController {
         @Override
         public User mapRow(ResultSet rs, int rowNum) throws SQLException {
             User user = new User();
-            user.setUid(rs.getInt("uid"));
+            user.setUid(rs.getLong("uid"));
             user.setEmail(rs.getString("email"));
             user.setInterest(rs.getString("interest"));
             return user;
